新增自动生成前端代码功能

+++ JMS.Common/Net/NetClient.cs
@@ -131,7 +131,7 @@ namespace JMS
                     {
                         this.WriteServiceData(new GatewayCommand()
                         {
-                            Type = CommandType.HealthyCheck
+                            Type = (int)CommandType.HealthyCheck
                         });
                     }
                     catch (Exception ex)
