+++ JMS.Gateway/Domains/MicroServiceReception.cs
@@ -127,7 +127,7 @@ namespace JMS.Domains
                 {
                     var command = await NetClient.ReadServiceObjectAsync<GatewayCommand>();
 
-                    if (command.Type == CommandType.ReportClientConnectQuantity)
+                    if (command.Type == (int)CommandType.ReportClientConnectQuantity)
                     {
                         //微服务向我报告当前它的请求连接数
                         //_Logger?.LogDebug($"微服务{this.ServiceInfo.ServiceNames.ToJsonString()} {this.ServiceInfo.Host}:{this.ServiceInfo.Port} 当前连接数：{command.Content}");
